Adds extensive documentation for package
The old tutorial notebook is extended and renamed to README.ipynb
Sphinx documentation is added and the notebook content is automatically
converted to rst so it can serve as the documentation start page.

Also includes a Makefile for building the documentation rst and html structure.
Requirement to build documentation:
* ipython nbconvert
* pandoc
* sphinx

Adds a submodule containing the gh-pages brach of the same repo
for easy publishing of the generated docs to github's pages