Use a relative submodules url to avoid protocol problems
If I install pyschema with:

 pip install git+https://github.com/spotify/pyschema.git

it will automatically try to update submodules, which is hardcoded to
the git protocol. This requires ssh access as well as a github account
with a public key configured.

Using a relative url like ../pyschema.git will use the same protocol
as the remote.

Using only . works for me and does not seem to go over the network at
all, but it is not documented in the git docs.